+++ .eslintrc.js
@@ -4,7 +4,7 @@ module.exports = {
     'plugin:@typescript-eslint/recommended-requiring-type-checking',
     'plugin:react/recommended',
     'prettier/@typescript-eslint',
-    'plugin:jest/recommended'
+    'plugin:jest/recommended',
     // 'plugin:jsx-a11y/recommended'
   ],
   parser: '@typescript-eslint/parser',
@@ -19,7 +19,7 @@ module.exports = {
     browser: true,
     node: true,
     es6: true,
-    "jest/globals": true
+    'jest/globals': true,
   },
   globals: {
     page: true,
