+++ src/common/dev/store.ts
@@ -1,3 +1,3 @@
-import { store } from '../../store';
+import { store } from '@store';
 
 export default store;
