+++ src/components/onboarding/data.ts
@@ -3,7 +3,7 @@ import {
   SECRET_KEY_FAQ_LOSE,
   SECRET_KEY_FAQ_WHEN,
   SECRET_KEY_FAQ_PASSWORD,
-} from '../../../common/track';
+} from '@common/track';
 
 const faqs = (appName: string) => {
   return [
