+++ src/components/onboarding/screens/connect.tsx
@@ -1,13 +1,13 @@
 import React, { useState, ChangeEvent } from 'react';
 import { Flex, Box, Text, Input } from '@blockstack/ui';
-import { AppIcon } from '../../app-icon';
-import { Link } from '../../../link';
-import { doTrack, CONNECT_SAVED, CONNECT_INCORRECT, CONNECT_BACK } from '../../../../common/track';
+import { AppIcon } from '@components/app-icon';
+import { Link } from '@components/link';
+import { doTrack, CONNECT_SAVED, CONNECT_INCORRECT, CONNECT_BACK } from '@common/track';
 import { useSelector } from 'react-redux';
-import { AppState } from '../../../../store';
-import { selectAppName, selectSecretKey } from '../../../../store/onboarding/selectors';
-import { ScreenHeader } from '../../header';
-import { ScreenBody, ScreenActions, ScreenFooter, Screen } from '../../screen';
+import { AppState } from '@store';
+import { selectAppName, selectSecretKey } from '@store/onboarding/selectors';
+import { ScreenHeader } from '@components/header';
+import { ScreenBody, ScreenActions, ScreenFooter, Screen } from '@components/screen';
 
 const ErrorText: React.FC = ({ children }) => (
   <Text textAlign="left" display="block" color="#de0014" mt={2}>
