+++ src/components/onboarding/screens/create.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useDispatch } from 'react-redux';
 import { Box, Spinner, Flex, Text } from '@blockstack/ui';
 
-import { doCreateSecretKey } from '../../../../store/onboarding/actions';
+import { doCreateSecretKey } from '@store/onboarding/actions';
 import { ScreenHeader } from '../../header';
 import { Screen, ScreenBody } from '../../screen';
