+++ src/components/onboarding/screens/decrypt-recovery-code.tsx
@@ -2,11 +2,11 @@ import React, { useState } from 'react';
 import { Box, Input, Text } from '@blockstack/ui';
 import { decrypt } from '@blockstack/keychain';
 import { useSelector, useDispatch } from 'react-redux';
-import { AppState } from '../../../../store';
-import { selectMagicRecoveryCode } from '../../../../store/onboarding/selectors';
-import { doTrack, SIGN_IN_CORRECT } from '../../../../common/track';
-import { doStoreSeed } from '../../../../store/wallet/actions';
-import { DEFAULT_PASSWORD } from '../../../../store/onboarding/types';
+import { AppState } from '@store';
+import { selectMagicRecoveryCode } from '@store/onboarding/selectors';
+import { doTrack, SIGN_IN_CORRECT } from '@common/track';
+import { doStoreSeed } from '@store/wallet/actions';
+import { DEFAULT_PASSWORD } from '@store/onboarding/types';
 import { ScreenHeader } from '../../header';
 import { Screen, ScreenBody, ScreenActions } from '../../screen';
