+++ src/components/onboarding/screens/index.tsx
