+++ src/components/onboarding/screens/save-key.tsx
@@ -2,12 +2,12 @@ import React from 'react';
 import { useSelector } from 'react-redux';
 
 import { Collapse } from '../../collapse';
-import { AppState } from '../../../../store';
+import { AppState } from '@store';
 
-import { selectAppName } from '../../../../store/onboarding/selectors';
+import { selectAppName } from '@store/onboarding/selectors';
 import { faqs } from '../data';
 
-import { doTrack, SECRET_KEY_INSTR_CONFIRMED } from '../../../../common/track';
+import { doTrack, SECRET_KEY_INSTR_CONFIRMED } from '@common/track';
 
 import { ScreenHeader } from '../../header';
 import { Screen, ScreenBody, ScreenActions } from '../../screen';
