+++ src/components/onboarding/screens/secret-key.tsx
@@ -5,10 +5,10 @@ import { Toast } from '../../toast';
 import { Card } from '../../card';
 import { SeedTextarea } from '../../seed-textarea';
 
-import { doTrack, SECRET_KEY_INTRO_COPIED } from '../../../../common/track';
+import { doTrack, SECRET_KEY_INTRO_COPIED } from '@common/track';
 
-import { AppState } from '../../../../store';
-import { selectSecretKey } from '../../../../store/onboarding/selectors';
+import { AppState } from '@store';
+import { selectSecretKey } from '@store/onboarding/selectors';
 import { ScreenHeader } from '../../header';
 import { Screen, ScreenBody, ScreenActions } from '../../screen';
