+++ src/components/window/index.tsx
